Get rid of some warnings under VS2003.
git-svn-id: svn://gil-w2k8/HexEdit_Repository/trunk@23 295070fd-a27b-da4f-a32e-77c6a882a0e3

+++ HexEdit/GridCtrl_src/GridCtrl.cpp
@@ -1328,8 +1328,8 @@ void CGridCtrl::OnEndInPlaceEdit(NMHDR* pNMHDR, LRESULT* pResult)
     case VK_PRIOR:
     case VK_HOME:
     case VK_END:
-        OnKeyDown(pgvItem->lParam, 0, 0);
-        OnEditCell(m_idCurrentCell.row, m_idCurrentCell.col, CPoint( -1, -1), pgvItem->lParam);
+        OnKeyDown(UINT(pgvItem->lParam), 0, 0);
+        OnEditCell(m_idCurrentCell.row, m_idCurrentCell.col, CPoint( -1, -1), UINT(pgvItem->lParam));
     }
 
     *pResult = 0;
